fix:add spotifycontext
